[LFC][IFC][Quirk] Let's adopt a less quirky behavior for line breaks on non-empty lines https://bugs.webkit.org/show_bug.cgi?id=220252
Reviewed by Antti Koivisto.

WebKit has this quirk behavior where <br> stretches the line box vertically only when the line is empty.
e.g.

<div><span><br></span></div>
We make the line as tall as the <br> is but

<div>some content<span style="font-size: 200px;"><br></span></div>
does not produce a ~200px tall line.

However when horizontal padding/border added, the line suddenly becomes ~200px tall:
<div>some content<span style="font-size: 200px; padding: 1px;"><br></span></div>

Removing this quirk makes the inlineLevelBoxAffectsLineBox logic simpler since we don't have to check against
whether the line is considered empty and the result is aligned with Chrome's behavior.
(This is in preparation for getting rid of the is-considered-empty concept for line/line box.)

* layout/inlineformatting/InlineFormattingContextGeometry.cpp:
(WebCore::Layout::LineBoxBuilder::constructInlineLevelBoxes):
* layout/inlineformatting/InlineFormattingContextQuirks.cpp:
(WebCore::Layout::InlineFormattingContext::Quirks::inlineLevelBoxAffectsLineBox const):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271481 268f45cc-cd09-0410-ab3c-d52691b4dbfc