Crash from CompositeEditCommand::moveParagraphs() being passed null end https://bugs.webkit.org/show_bug.cgi?id=220630
Patch by Julian Gonzalez <julian_a_gonzalez@apple.com> on 2021-01-14
Reviewed by Ryosuke Niwa.

Source/WebCore:

If the start or end VisiblePositions inside InsertListCommand::unlistifyParagraph()
are null VisiblePositions (even if they are not null Positions), passing them to the
call to moveParagraphs() at the end of the function isn't meaningful and will result
in a crash - so return early in this case.

Test: editing/inserting/paragraph-outdent-crash.html

* editing/InsertListCommand.cpp:
(WebCore::InsertListCommand::unlistifyParagraph):

LayoutTests:

Add a test to verify that the crash here is resolved, and also remove a newline
from another related test that now renders one fewer newline
(in that case, the ultimate bug is similar, so the result should be similar).

* editing/inserting/insert-list-in-iframe-in-list-expected.txt:
* editing/inserting/paragraph-outdent-crash-expected.txt: Added.
* editing/inserting/paragraph-outdent-crash.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271510 268f45cc-cd09-0410-ab3c-d52691b4dbfc