REGRESSION(r269813): PLT5 regressed by 0.5% https://bugs.webkit.org/show_bug.cgi?id=220652 <rdar://problem/71517335>
Reviewed by Antti Koivisto.

In r269813 we added support for animating more pseudo-elements besides ::after and ::before. However, the only other
pseudo-element we realistically should be supporting animations for at this juncture is ::marker. So instead of
calling Style::TreeResolver::resolvePseudoStyle() for all public pseudo-elements, we call it for ::after, ::before
and ::marker alone.

* style/StyleTreeResolver.cpp:
(WebCore::Style::TreeResolver::resolveElement):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271515 268f45cc-cd09-0410-ab3c-d52691b4dbfc