When non-integer tabindex is set, the behavior of element should be same as the tabindex is omitted. https://bugs.webkit.org/show_bug.cgi?id=220648
Reviewed by Antti Koivisto.

Source/WebCore:

When non-integer tabindex is specified, the element should behave the same way as the tabindex is omitted.
https://html.spec.whatwg.org/multipage/interaction.html#attr-tabindex

WebKit didn't overwrite the internal tabindex value when the new value is non-integer.

Test: LayoutTests\fast\html\tabindex-overwrite-with-non-integer.html

* html/HTMLElement.cpp:
(WebCore::HTMLElement::parseAttribute): If the new value cannot be parsed as the integer, clear the existing tabindex.

LayoutTests:

Add LayoutTest case for tabindex which is overwritten by non-integers.
When non-integer tabindex is specified, the element should behave the same way as the tabindex is omitted.
https://html.spec.whatwg.org/multipage/interaction.html#attr-tabindex

* fast/html/tabindex-overwrite-with-non-integer-expected.txt: Added.
* fast/html/tabindex-overwrite-with-non-integer.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271523 268f45cc-cd09-0410-ab3c-d52691b4dbfc