createImageBitmap() causes null pointer dereference using Cairo lib https://bugs.webkit.org/show_bug.cgi?id=220584
Reviewed by Carlos Garcia Campos.

Source/WebCore:

createImageBitmap(ImageData) creates an internal ImageBuffer with
the same size of the given ImageData. Cairo port has the
limitation to maximum width/height of ImageBuffer as 32768. It
fails to create the internal ImageBuffer if the ImageData size is
greater than 32768. Added a null-checking for the case.

Test: fast/images/createImageBitmap-with-ImageData.html

* html/ImageBitmap.cpp:
(WebCore::ImageBitmap::createPromise): Added a null-checking for
tempBitmapData. Call resolveWithBlankImageBuffer in the case.
* platform/graphics/cairo/ImageBufferCairoImageSurfaceBackend.cpp:
(WebCore::ImageBufferCairoImageSurfaceBackend::create): Check the
cairo_surface_status of 'surface' to avoid the assertion failure
in ImageBufferCairoImageSurfaceBackend ctor.

LayoutTests:

* fast/images/createImageBitmap-with-ImageData-expected.txt: Added.
* fast/images/createImageBitmap-with-ImageData.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271583 268f45cc-cd09-0410-ab3c-d52691b4dbfc