REGRESSION(r266695) Range control with custom track width sized incorrectly https://bugs.webkit.org/show_bug.cgi?id=220608
Reviewed by Zalan Bujtas.

Source/WebCore:

Test: fast/forms/range/input-range-sizing-fixed-size.html

In r266695 we modified the way min-{width|height}: auto was computed for flexbox elements. That broke the sizing
of input range controls that were working under the assumption that min-{width|height} was 0. We have to force it
in the UA CSS in order to keep the same behaviour.

* css/html.css:
(input[type="range"]::-webkit-slider-container, input[type="range"]::-webkit-media-slider-container): force
min-width: 0.

LayoutTests:

New test case to verify that input range controls are sized correctly with specified widths.

* fast/forms/range/input-range-sizing-fixed-size-expected.html: Added.
* fast/forms/range/input-range-sizing-fixed-size.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271618 268f45cc-cd09-0410-ab3c-d52691b4dbfc