A DOMWindow should not be reused if its JSDOMWindow has been created https://bugs.webkit.org/show_bug.cgi?id=220658 <rdar://problem/70335075>
Reviewed by Geoffrey Garen.

Source/WebCore:

When a JSDOMWindow is created from a DOMWindow, its document is used to decide whether to expose some properties,
based on SecureContext or quirks.
It is important to ensure that the document has the correct information at that time.

Add a boolean to DOMWindow to identify whether it was wrapped without its document being properly initialized.
Use that boolean in DocumentWriter to forbidd reusing the DOMWindow if it was created this way.

Covered by API test.

* bindings/js/JSWindowProxy.cpp:
(WebCore::JSWindowProxy::setWindow):
* dom/SecurityContext.h:
(WebCore::SecurityContext::haveInitializedSecurityOrigin const):
(WebCore::SecurityContext::didFailToInitializeSecurityOrigin):
* loader/DocumentWriter.cpp:
(WebCore::DocumentWriter::begin):
* page/DOMWindow.h:

Tools:

* TestWebKitAPI/Tests/WebKitCocoa/WKWebViewEvaluateJavaScript.mm:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271642 268f45cc-cd09-0410-ab3c-d52691b4dbfc