[WPE] focus-visible-003.html and focus-visible-004.html are timing out in input tags https://bugs.webkit.org/show_bug.cgi?id=220578
Reviewed by Javier Fernandez.

Source/WebCore:

Dispatch "focus" event for form controls when they are focused via mouse click.
This was already done in GTK+ but was not yet in WPE.

This works different on Mac ports as it's a particularity of how focus works on Mac platforms
(see https://webkit.org/b/22261 for details).

* html/HTMLFormControlElement.cpp:
(WebCore::HTMLFormControlElement::isMouseFocusable const): Add PLATFORM(WPE).

LayoutTests:

* platform/wpe/TestExpectations: Unskip tests that are not failing anymore.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271671 268f45cc-cd09-0410-ab3c-d52691b4dbfc