Update font when resolving letter-spacing: calc(...) values https://bugs.webkit.org/show_bug.cgi?id=216659
Patch by Frederic Wang <fwang@igalia.com> on 2021-01-21
Reviewed by Ryosuke Niwa.

Source/WebCore:

In bug 176215, maybeUpdateFontForLetterSpacing was added to ensure a font is available for
resolution of letter-spacing values that rely on relative font lengths. However, this does
not take into account the case when these lengths are part of a calc expression. In order to
keep things simple, this patch unconditionally updates the font when letter-spacing is a
calc expression.

Test: fast/css/letter-spacing-calc-with-font-relative-lengths-crash.html

* style/StyleBuilderCustom.h:
(WebCore::Style::maybeUpdateFontForLetterSpacing):

LayoutTests:

* fast/css/letter-spacing-calc-with-font-relative-lengths-crash-expected.txt: Added.
* fast/css/letter-spacing-calc-with-font-relative-lengths-crash.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271688 268f45cc-cd09-0410-ab3c-d52691b4dbfc