Null dereference in CompositeEditCommand::cloneParagraphUnderNewElement(): needs to check lastNode parent https://bugs.webkit.org/show_bug.cgi?id=220813
Patch by Julian Gonzalez <julian_a_gonzalez@apple.com> on 2021-01-21
Reviewed by Ryosuke Niwa.

Source/WebCore:

When iterating through parent nodes, cloneParagraphUnderNewElement()
has to be careful to check for the top of the DOM tree for *both*
startNode and lastNode.
Since we're explicitly checking for this now, remove the assertion
that lastNode's parent is never nullptr.

Test: editing/inserting/insert-ul-select-all.html

* editing/CompositeEditCommand.cpp:
(WebCore::CompositeEditCommand::cloneParagraphUnderNewElement):

LayoutTests:

Add a test to verify that we don't iterate outside the DOM tree when
inserting an unordered list on top of existing content.

* editing/inserting/insert-ul-select-all-expected.txt: Added.
* editing/inserting/insert-ul-select-all.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271720 268f45cc-cd09-0410-ab3c-d52691b4dbfc