Denying storage access on playstation.com then clicking sign-in fails to show the prompt again https://bugs.webkit.org/show_bug.cgi?id=220838 <rdar://problem/73477444>
Reviewed by Alex Christensen.

Temporary cross-page storage access in the Web Process for quirks
should only be granted if the user hits "Allow" on the prompt.
This was causing the prompt to fail to show if a user clicks "Don't
Allow" then tries to sign in, then sign-in would fail without
storage access in the Network Process.

* page/Quirks.cpp:
(WebCore::Quirks::requestStorageAccessAndHandleClick const):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271744 268f45cc-cd09-0410-ab3c-d52691b4dbfc